banktransactions/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
python-dotenv>=0.15.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
flask-login==0.6.3
pytest==7.4.4
pytest-cov==4.1.0
certifi==2024.2.2
gunicorn==21.2.0
google-genai>=1.14.0
requests>=2.25.0
redis==5.0.1
rq==1.15.1
rq-scheduler==0.13.1
cryptography==42.0.2
python-jose==3.3.0